Mood Detection Improvements
Removed the values for Fear and Surprise and weighted the remaining values against each other. Provided more readable output.
