+++ Text2Emotion/Extra/__pycache__/text2emotion.cpython-310.pyc
