+++ Text2Emotion/Extra/testing2.py
