--- Text2Emotion/Lyrics_DontStopBelieving2.txt
@@ -1,1 +1,0 @@
-Just a small town girl Living in a lonely world She took the midnight train going anywhere Just a city boy Born and raised in South Detroit He took the midnight train going anywhere A singer in a smokey room A smell of wine and cheap perfume For a smile they can share the night It goes on and on and on and on Strangers waiting Up and down the boulevard Their shadows searching in the night Streetlights people Living just to find emotion Hiding somewhere in the night Working hard to get my fill Everybody wants a thrill Paying anything to roll the dice Just one more time Some will win some will lose Some are born to sing the blues Whoa the movie never ends It goes on and on and on and on Strangers waitin Up and down the boulevard Their shadows searching in the night Streetlights people Living just to find emotion Hiding somewhere in the night Do not stop believing Hold on to that feeling Streetlights people Do not stop believing Hold on Streetlightsg people Do not stop believing Hold on to that feeling Streetlights people
